Overhaul larascript.sh
Separate app name and domain name questions.
Domain name suggestion.
Check existence public folder.
Use alias, service provider and composer helpers.
Dynamically load packages.

+++ larascript.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 
-# Get the source directory
-DIR="$( cd "$( echo "${BASH_SOURCE[0]%/*}" )"; pwd )"
+#
+# Larascript. Bash interactive Laravel setup script.
+# https://github.com/sq2/larascript
+# By @Codepl
+#
+
+
+# Set the source path
+SOURCE_PATH="$( cd "$( echo "${BASH_SOURCE[0]%/*}" )"; pwd )"
+
+# Include functions
+. "$SOURCE_PATH"/helpers/init.sh
 
 
 #-----------------------------------------------------------------------
@@ -12,28 +22,33 @@ echo
 echo "Get Started Configuring a New Laravel Website"
 echo
 
-# App name?
+# App name
 echo
-echo -n "What is the local domain name of this app, without the extension? : "
+echo -n "App name? (Maybe the domain name without the extension) : "
 read appname
-domain="$appname.dev"
+
+# Local domain name
+echo
+read -p "Local domain name? ["$appname.dev"] : " domain
+domain=${domain:-"$appname.dev"}
 
 # Create new laravel Project
 echo
 echo -n "Create a new Laravel app? (y/n) : "
 read -e laravel
 if [[ $laravel == "y" ]]
-    then
-        # Assumes that laravel.phar is available globally.
-        # wget http://laravel. com/laravel.phar
-        # chmod 755 laravel.phar
-        # mv laravel.phar /usr/local/bin/laravel
-        laravel new $domain
+then
+    # Assumes that laravel.phar is available globally.
+    # wget http://laravel. com/laravel.phar
+    # chmod 755 laravel.phar
+    # mv laravel.phar /usr/local/bin/laravel
+    laravel new $domain
 
-        # Use create-project instead
-        # composer create-project laravel/laravel $domain --prefer-dist
+    # Use create-project instead
+    # composer create-project laravel/laravel $domain --prefer-dist
 
-        cd $domain
+    cd $domain
+    WORK_PATH=$(pwd)
 fi
 
 
@@ -42,29 +57,31 @@ fi
 #-----------------------------------------------------------------------
 
 # Move public files to root. Shit happens.
-echo
-echo -n "Move public files to root for shared hosting? (y/n) : "
-read -e public
-if [[ $public == "y" ]]
-then
-    # Move files and cleanup
-    cd public; mv * ../; cd ..
-    rm -rf public; rm readme.md; rm CONTRIBUTING.md
+if [ -d "public" ]; then
+    echo
+    echo -n "Move public files to root for shared hosting? (y/n) : "
+    read -e public
+    if [[ $public == "y" ]]
+    then
+        # Move files and cleanup
+        cd public; mv * ../; cd ..
+        rm -rf public; rm readme.md; rm CONTRIBUTING.md
 
-    # Fix paths
-    gsed -i "s@/../bootstrap@/bootstrap@g" index.php
-    gsed -i "s@/../public@/..@" bootstrap/paths.php
+        # Fix paths
+        gsed -i "s@/../bootstrap@/bootstrap@g" index.php
+        gsed -i "s@/../public@/..@" bootstrap/paths.php
 
-    # Copy .htaccess file
-    cp $DIR/src/public/.htaccess .
-else
-    # Cleanup. Should probably be done in Git ignore file instead.
-    cd public
-    rm readme.md; rm CONTRIBUTING.md
-    cd ..
+        # Copy .htaccess file
+        cp $SOURCE_PATH/src/public/.htaccess .
+    else
+        # Cleanup. Should probably be done in Git ignore file instead.
+        cd public
+        rm readme.md; rm CONTRIBUTING.md
+        cd ..
 
-    # Copy .htaccess file to public
-    cp $DIR/src/public/.htaccess public
+        # Copy .htaccess file to public
+        cp $SOURCE_PATH/src/public/.htaccess public
+    fi
 fi
 
 
@@ -96,21 +113,21 @@ echo
 echo -n "Does your app require a MySql database? : (y/n) "
 read -e mysqldb
 if [[ $mysqldb == 'y' ]]
-    then
-        echo -n "What is the name of the database for this app? : "
-        read -e database
+then
+    echo -n "What is the name of the database for this app? : "
+    read -e database
 
-        echo -n "Enter a database password? : "
-        read -e password
+    echo -n "Enter a database password? : "
+    read -e password
 
-        echo "Creating MySQL database"
-        echo "Enter system password"
-        sudo mysql -uroot -p$password -e"CREATE DATABASE $database"
+    echo "Creating MySQL database"
+    echo "Enter system password"
+    sudo mysql -uroot -p$password -e"CREATE DATABASE $database"
 
-        echo Updating database configuration file
-        gsed -i "s/'database' => 'database'/'database' => '$database'/" app/config/database.php
-        gsed -i "s/'password'  => ''/'password'  => '$password'/" app/config/database.php
-        # gsed -i "s/'username'  => 'root'/'username'  => '$username'/" app/config/database.php
+    echo Updating database configuration file
+    gsed -i "s/'database' => 'database'/'database' => '$database'/" app/config/database.php
+    gsed -i "s/'password'  => ''/'password'  => '$password'/" app/config/database.php
+    # gsed -i "s/'username'  => 'root'/'username'  => '$username'/" app/config/database.php
 fi
 
 
@@ -123,61 +140,61 @@ echo
 echo -n "Add custom libraries and settings to $domain? (y/n) : "
 read -e custom
 if [[ $custom == "y" ]]
-    then
-        # Session settings
-        echo
-        echo "Configuring settings..."
-        gsed -i "s/'lifetime' => 120/'lifetime' => 240/" app/config/session.php
-        gsed -i "s/'cookie' => 'laravel_session'/'cookie' => '$appname_session'/" app/config/session.php
+then
+    # Session settings
+    echo
+    echo "Configuring settings..."
+    gsed -i "s/'lifetime' => 120/'lifetime' => 240/" app/config/session.php
+    gsed -i "s/'cookie' => 'laravel_session'/'cookie' => '$appname_session'/" app/config/session.php
 
-        # Cache settings
-        gsed -i "s/'prefix' => 'laravel'/'prefix' => '$appname'/" app/config/cache.php
+    # Cache settings
+    gsed -i "s/'prefix' => 'laravel'/'prefix' => '$appname'/" app/config/cache.php
 
-        # Workbench settings
-        # gsed -i "s/'name' => ''/'name' => 'Your Name'/" app/config/workbench.php
-        # gsed -i "s/'email' => ''/'email' => 'Your Email Address'/" app/config/workbench.php
+    # Workbench settings
+    # gsed -i "s/'name' => ''/'name' => 'Your Name'/" app/config/workbench.php
+    # gsed -i "s/'email' => ''/'email' => 'Your Email Address'/" app/config/workbench.php
 
-        # Append to global.php file
-        cat "$DIR/src/app/start/global.php" >> app/start/global.php
+    # Append to global.php file
+    cat "$SOURCE_PATH/src/app/start/global.php" >> app/start/global.php
 
-        # Add extra files for easier management.
-        printf "<?php\n\n// View composers" > app/composers.php
+    # Add extra files for easier management.
+    printf "<?php\n\n// View composers" > app/composers.php
 
-        # Views
-        mkdir app/views/layouts
-        mkdir app/views/auth
-        mkdir app/views/errors
+    # Views
+    mkdir app/views/layouts
+    mkdir app/views/auth
+    mkdir app/views/errors
 
-        # Add asset folders
-        echo "Adding asset folders..."
-        mkdir img
-        mkdir includes
-        mkdir javascript
-        mkdir less
+    # Add asset folders
+    echo "Adding asset folders..."
+    mkdir img
+    mkdir includes
+    mkdir javascript
+    mkdir less
 
-        # Copy library folders
-        # echo "Copying library folders..."
-        # cp -R $DIR/lib app/
+    # Copy library folders
+    # echo "Copying library folders..."
+    # cp -R $SOURCE_PATH/lib app/
 
-        # Add service providers
-        # echo "Adding service providers..."
-        # gsed -i "/WorkbenchServiceProvider/a \\\t\t'VendorName\\\Product\\\ProductServiceProvider'," app/config/app.php
+    # Add service providers
+    # echo "Adding service providers..."
+    # add_service_provider "VendorName\Product\ProductServiceProvider"
 
-        # Add facades
-        echo "Adding facades..."
-        gsed -i "/'View'/a \\\t\t'Carbon' => 'Carbon\\\Carbon'," app/config/app.php
+    # Add aliases
+    echo "Adding facade aliases..."
+    add_alias "Carbon" "Carbon\Carbon"
 
-        # Add psr-0 entries
-        # echo "Adding psr-0 entries..."
-        # php $DIR/helpers/add_to_json.php composer.json .autoload.psr-0.Helpers.app/lib
+    # Add psr-0 entries
+    # echo "Adding psr-0 entries..."
+    # add_to_composer ".autoload.psr-0.Helpers.app/lib"
 
-        # Add psr-4 entries
-        # echo "Adding psr-4 entries..."
-        # php $DIR/helpers/add_to_json.php composer.json ".autoload.psr-4.Helpers\\.app/lib"
+    # Add psr-4 entries
+    # echo "Adding psr-4 entries..."
+    # add_to_composer ".autoload.psr-4.Helpers\\.app/lib"
 
-        # Add to classmap
-        # echo "Adding classmap entries..."
-        # php $DIR/helpers/add_to_json.php composer.json ".autoload.classmap.app/composers" array
+    # Add to classmap
+    # echo "Adding classmap entries..."
+    # add_to_composer ".autoload.classmap.app/composers" array
 fi
 
 
@@ -185,39 +202,27 @@ fi
 # PACKAGES                                                             |
 #-----------------------------------------------------------------------
 
-# Add Way/Generators package to composer.json
 echo
-echo -n "Add Way/Generators package to $domain? (y/n) : "
-read -e generators
-if [[ $generators == "y" ]]
+for f in "$SOURCE_PATH"/packages/*.sh; do
+    filename=$(basename $f)
+    package=${filename%.*}
+    package=$(echo $package | tr "_" " ")
+
+    ask="packagedev"
+
+    echo -n "Add $package package? (y/n) : "
+    read -e ask
+    if [[ $ask == "y" ]]
     then
-        echo "Adding Way/Generators to $domain..."
+        echo
+        echo "Adding $package..."
+        echo
 
-        composer require --dev --no-update way/generators:dev-master
-
-        # Add service provider
-        gsed -i "/WorkbenchServiceProvider/a \\\t\t'Way\\\Generators\\\GeneratorsServiceProvider'," app/config/app.php
-fi
-
-# Add Clockwork package to composer.json. Recommended by Jeffrey Way at laracasts.com.
-echo
-echo -n "Add itsgoingd/clockwork package to $domain? (y/n) : "
-read -e clockwork
-if [[ $clockwork == "y" ]]
-    then
-        echo "Adding Clockwork to $domain..."
-
-        composer require --dev --no-update itsgoingd/clockwork:dev-master
-
-        # Add service provider
-        gsed -i "/WorkbenchServiceProvider/a \\\t\t'Clockwork\\\Support\\\Laravel\\\ClockworkServiceProvider'," app/config/app.php
-
-        # Add facade
-        gsed -i "/'View'/a \\\t\t'Clockwork' => 'Clockwork\\\Support\\\Laravel\\\Facade'," app/config/app.php
-
-        # Append to local.php file
-        cat "$DIR/src/vendor/clockwork.php" >> app/start/local.php
-fi
+        . "$f"
+    else
+        echo
+    fi
+done
 
 
 #-----------------------------------------------------------------------
@@ -229,8 +234,8 @@ echo
 echo -n "Run Composer update? (y/n) : "
 read -e composer
 if [[ $composer == "y" ]]
-    then
-        composer update
+then
+    composer update
 fi
 
 # What else?
@@ -239,6 +244,7 @@ echo "----------------------------------------------------------------"
 echo
 echo "The following items will need to be handled manually (for now):"
 echo
-echo "Bring in Javascript and CSS assets."
+echo "Bring in Javascript, CSS and image assets."
 echo "Error handling for missing pages."
+echo "Setup virtual host and local domain. (Coming soon)"
 echo
